+++ strokepred/api/__init__.py
@@ -1,0 +1,1 @@
+from strokepred.api.app import app
