+++ strokepred/base/dataset.py
